refactor(runtime): split methods_signature.rs into cohesive modules (§7-8)
`methods_signature.rs` was 1237 lines, over the 500-line module limit. Split it
into four files along its natural seams, each under 500 lines, with no behavior
change (the only fn-signature delta is `assumed_signature_param_defs` going
private→`pub(super)`, now called from the candidates module):

- `methods_signature.rs` (450) — shape extraction from defaults
  (`extract_shape_from_default`, `extract_dims_from_shape_expr`),
  `coerce_attr_value_by_sigil`, placeholder positional/named auto-signature
  detection (`auto_signature_uses`), and `assumed_signature_param_defs`.
- `methods_signature_errors.rs` (180) — the structured dispatch-error
  constructors (`make_private_permission_error`,
  `make_private_unqualified_error`, `make_method_not_found_error`,
  `make_x_immutable_error`, `make_multi_no_match_error`,
  `make_multi_ambiguous_error`, `make_multi_no_match_error_detailed`).
- `methods_signature_candidates.rs` (348) — named-param-key collection,
  call-arg candidate matching, candidate routine lookup, signature value
  building, and arity/count Value builders.
- `methods_signature_shaped.rs` (274) — shaped-array dim inference/construction
  (`shaped_dims_from_new_args`, `make_shaped_array`, `infer_array_shape`),
  parametric-type-name parsing, and multi-candidate `call_method_all_with_values`.

`methods_signature_errors` is declared `pub(crate) mod` (its `pub(crate)`
`make_multi_ambiguous_error` is used from the VM). Each new file carries only
the imports it uses (`Symbol` → errors, `CallArg` → kept, `value::signature`
builders → candidates); call sites across the codebase are updated to the
`methods_signature_errors::` path. Fifth slice of the §7-8 file-splitting sweep.
