+++ src/runtime/methods.rs
@@ -1,4 +1,4 @@
-use super::methods_signature::{
+use super::methods_signature_errors::{
     make_method_not_found_error, make_multi_no_match_error, make_x_immutable_error,
 };
 use super::*;
