+++ src/runtime/methods_signature.rs
@@ -1,183 +1,5 @@
 use super::*;
 use crate::ast::CallArg;
-use crate::symbol::Symbol;
-use crate::value::signature::{make_signature_value_with_owner, param_defs_to_sig_info};
-
-/// Create a structured X::Method::Private::Permission error: a fully-qualified
-/// private call `$o!Owner::meth` where `Owner` does not trust the calling
-/// package. Carries `method`, `source-package` (the owner) and `calling-package`.
-pub(super) fn make_private_permission_error(
-    method_name: &str,
-    class_name: &str,
-    calling_package: &str,
-) -> RuntimeError {
-    let msg = format!(
-        "Cannot call private method '{}' on package {} because it does not trust {}",
-        method_name, class_name, calling_package
-    );
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("method".to_string(), Value::str(method_name.to_string()));
-    attrs.insert(
-        "source-package".to_string(),
-        Value::str(class_name.to_string()),
-    );
-    attrs.insert(
-        "calling-package".to_string(),
-        Value::str(calling_package.to_string()),
-    );
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Method::Private::Permission"), attrs);
-    let mut err = RuntimeError::new(&msg);
-    err.exception = Some(Box::new(ex));
-    err
-}
-
-/// Create a structured X::Method::Private::Unqualified error: an unqualified
-/// private call `$o!meth` on something other than `self`. Raku requires such
-/// calls to name the package that defines the private method.
-pub(super) fn make_private_unqualified_error(method_name: &str) -> RuntimeError {
-    let msg = format!(
-        "Calling private method '{}' must be fully qualified with the package containing that private method.",
-        method_name
-    );
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("method".to_string(), Value::str(method_name.to_string()));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Method::Private::Unqualified"), attrs);
-    let mut err = RuntimeError::new(&msg);
-    err.exception = Some(Box::new(ex));
-    err
-}
-
-/// Create a structured X::Method::NotFound error.
-pub(super) fn make_method_not_found_error(
-    method_name: &str,
-    type_name: &str,
-    private: bool,
-) -> RuntimeError {
-    use super::did_you_mean::{known_methods_for_type, suggest_method};
-
-    let mut msg = if private {
-        format!(
-            "No such private method '{}' for invocant of type '{}'",
-            method_name, type_name
-        )
-    } else {
-        format!(
-            "No such method '{}' for invocant of type '{}'",
-            method_name, type_name
-        )
-    };
-
-    // Append "Did you mean ...?" suggestion if a close match exists
-    if !private {
-        let candidates = known_methods_for_type(type_name);
-        if let Some(suggestion) = suggest_method(method_name, candidates) {
-            msg.push_str(&format!("\nDid you mean '{}'?", suggestion));
-        }
-    }
-
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("method".to_string(), Value::str(method_name.to_string()));
-    attrs.insert("typename".to_string(), Value::str(type_name.to_string()));
-    attrs.insert("private".to_string(), Value::Bool(private));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Method::NotFound"), attrs);
-    let mut err = RuntimeError::new(&msg);
-    err.exception = Some(Box::new(ex));
-    err
-}
-
-/// Create a structured X::Immutable error.
-pub(super) fn make_x_immutable_error(method_name: &str, typename: &str) -> RuntimeError {
-    let msg = format!(
-        "Cannot call '{}' on an immutable '{}'",
-        method_name, typename
-    );
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("method".to_string(), Value::str(method_name.to_string()));
-    attrs.insert("typename".to_string(), Value::str(typename.to_string()));
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Immutable"), attrs);
-    let mut err = RuntimeError::new(&msg);
-    err.exception = Some(Box::new(ex));
-    err
-}
-
-/// Create a structured X::Multi::NoMatch error.
-pub(super) fn make_multi_no_match_error(method_name: &str) -> RuntimeError {
-    let msg = format!("No matching candidates for method: {}", method_name);
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Multi::NoMatch"), attrs);
-    let mut err = RuntimeError::new(&msg);
-    err.exception = Some(Box::new(ex));
-    err
-}
-
-/// Build an `X::Multi::Ambiguous` error naming the dispatch target and the
-/// candidate signatures that all matched equally well.
-pub(crate) fn make_multi_ambiguous_error(
-    method_name: &str,
-    invocant_type: &str,
-    candidate_sigs: &[String],
-) -> RuntimeError {
-    let sig_lines = if candidate_sigs.is_empty() {
-        String::new()
-    } else {
-        let mut s = String::new();
-        for sig in candidate_sigs {
-            s.push_str("\n  ");
-            s.push_str(sig);
-        }
-        s
-    };
-    let msg = format!(
-        "Ambiguous call to '{}({}: )'; these signatures all match:{}",
-        method_name, invocant_type, sig_lines
-    );
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Multi::Ambiguous"), attrs);
-    let mut err = RuntimeError::new(&msg);
-    err.exception = Some(Box::new(ex));
-    err
-}
-
-/// Build an `X::Multi::NoMatch` error with a Raku-style message that names the
-/// dispatch target (invocant type + method name) and lists the candidate
-/// signatures, e.g.:
-///
-/// ```text
-/// Cannot resolve caller has_tie(WorkingTie:D: Array:D); none of these
-/// signatures matches: ...
-/// ```
-pub(super) fn make_multi_no_match_error_detailed(
-    method_name: &str,
-    invocant_type: &str,
-    candidate_sigs: &[String],
-) -> RuntimeError {
-    let sig_lines = if candidate_sigs.is_empty() {
-        String::new()
-    } else {
-        let mut s = String::new();
-        for sig in candidate_sigs {
-            s.push_str("\n    ");
-            s.push_str(sig);
-        }
-        s
-    };
-    let msg = format!(
-        "Cannot resolve caller {}({}:D: ); none of these signatures matches:{}",
-        method_name, invocant_type, sig_lines
-    );
-    let mut attrs = std::collections::HashMap::new();
-    attrs.insert("message".to_string(), Value::str(msg.clone()));
-    let ex = Value::make_instance(Symbol::intern("X::Multi::NoMatch"), attrs);
-    let mut err = RuntimeError::new(&msg);
-    err.exception = Some(Box::new(ex));
-    err
-}
 
 impl Interpreter {
     /// Extract shape dimensions from a default expression that matches the pattern
@@ -546,7 +368,7 @@ impl Interpreter {
         (positional, named)
     }
 
-    fn assumed_signature_param_defs(
+    pub(super) fn assumed_signature_param_defs(
         data: &crate::value::SubData,
         assumed_positional: &[Value],
         assumed_named: &std::collections::HashMap<String, Value>,
@@ -624,614 +446,5 @@ impl Interpreter {
             }
         }
         Some(param_defs)
-    }
-
-    pub(super) fn collect_named_param_keys(
-        param_defs: &[ParamDef],
-        out: &mut std::collections::HashSet<String>,
-    ) {
-        for pd in param_defs {
-            if pd.named {
-                if pd.name == "__subsig__" {
-                    if let Some(key) = &pd.type_constraint {
-                        out.insert(key.clone());
-                    }
-                } else {
-                    out.insert(pd.name.clone());
-                }
-            }
-            if let Some(sub) = &pd.sub_signature {
-                Self::collect_named_param_keys(sub, out);
-            }
-        }
-    }
-
-    pub(super) fn has_named_slurpy_param(param_defs: &[ParamDef]) -> bool {
-        for pd in param_defs {
-            if pd.slurpy && pd.name.starts_with('%') {
-                return true;
-            }
-            if let Some(sub) = &pd.sub_signature
-                && Self::has_named_slurpy_param(sub)
-            {
-                return true;
-            }
-        }
-        false
-    }
-
-    pub(super) fn capture_to_call_args(value: &Value) -> Vec<Value> {
-        match value {
-            Value::Capture { positional, named } => {
-                let mut args = (**positional).clone();
-                for (k, v) in named.iter() {
-                    args.push(Value::Pair(k.clone(), Box::new(v.clone())));
-                }
-                args
-            }
-            other => vec![other.clone()],
-        }
-    }
-
-    pub(super) fn varref_parts(value: &Value) -> Option<(String, Value)> {
-        if let Value::Capture { positional, named } = value
-            && positional.is_empty()
-            && let Some(Value::Str(name)) = named.get("__mutsu_varref_name")
-            && let Some(inner) = named.get("__mutsu_varref_value")
-        {
-            return Some((name.to_string(), inner.clone()));
-        }
-        None
-    }
-
-    pub(super) fn var_target_from_meta_value(value: &Value) -> Option<String> {
-        match value {
-            Value::Mixin(inner, _) => Self::var_target_from_meta_value(inner),
-            Value::Instance { attributes, .. } => {
-                match attributes.as_map().get("__mutsu_var_target") {
-                    Some(Value::Str(name)) => Some(name.to_string()),
-                    _ => None,
-                }
-            }
-            _ => None,
-        }
-    }
-
-    pub(super) fn candidate_matches_call_args(
-        &mut self,
-        candidate: &Value,
-        args: &[Value],
-    ) -> bool {
-        match candidate {
-            Value::Sub(data) => {
-                if data.empty_sig && !args.is_empty() {
-                    return false;
-                }
-                if data.param_defs.is_empty() && !data.params.is_empty() {
-                    if args.iter().any(|arg| {
-                        matches!(
-                            arg,
-                            Value::Pair(key, _) if key != "__mutsu_test_callsite_line"
-                        )
-                    }) {
-                        return false;
-                    }
-                    let positional = args
-                        .iter()
-                        .filter(|arg| !matches!(arg, Value::Pair(_, _) | Value::ValuePair(_, _)))
-                        .count();
-                    return positional == data.params.len();
-                }
-                self.method_args_match(args, &data.param_defs)
-            }
-            Value::WeakSub(weak) => weak
-                .upgrade()
-                .is_some_and(|strong| self.method_args_match(args, &strong.param_defs)),
-            Value::Routine { name, .. } => self
-                .resolve_function_with_types(&name.resolve(), args)
-                .is_some(),
-            _ => false,
-        }
-    }
-
-    pub(super) fn routine_candidate_subs(&self, package: &str, name: &str) -> Vec<Value> {
-        let exact_local = format!("{package}::{name}");
-        let exact_global = format!("GLOBAL::{name}");
-        let prefix_local = format!("{package}::{name}/");
-        let prefix_global = format!("GLOBAL::{name}/");
-        let mut seen = std::collections::HashSet::new();
-        let mut out = Vec::new();
-        let mut multi_idx = 0usize;
-        for (key, def) in &self.registry().functions {
-            let key_s = key.resolve();
-            if key_s == exact_local
-                || key_s == exact_global
-                || key_s.starts_with(&prefix_local)
-                || key_s.starts_with(&prefix_global)
-            {
-                let fp =
-                    crate::ast::function_body_fingerprint(&def.params, &def.param_defs, &def.body);
-                if seen.insert(fp) {
-                    let mut env = self.env.clone();
-                    // Store the multi index for doc comment lookup
-                    env.insert(
-                        "__mutsu_multi_index".to_string(),
-                        Value::Int(multi_idx as i64),
-                    );
-                    out.push(Value::make_sub(
-                        def.package,
-                        def.name,
-                        def.params.clone(),
-                        def.param_defs.clone(),
-                        def.body.clone(),
-                        def.is_rw,
-                        env,
-                    ));
-                    multi_idx += 1;
-                }
-            }
-        }
-        out
-    }
-
-    pub(super) fn sub_signature_value(&self, data: &crate::value::SubData) -> Value {
-        let param_defs =
-            Self::assumed_signature_param_defs(data, &data.assumed_positional, &data.assumed_named)
-                .unwrap_or_else(|| {
-                    if !data.params.is_empty() {
-                        data.params
-                            .iter()
-                            .map(|name| ParamDef {
-                                name: name.clone(),
-                                default: None,
-                                multi_invocant: true,
-                                required: false,
-                                named: false,
-                                slurpy: false,
-                                double_slurpy: false,
-                                onearg: false,
-                                sigilless: false,
-                                type_constraint: None,
-                                literal_value: None,
-                                sub_signature: None,
-                                where_constraint: None,
-                                traits: Vec::new(),
-                                optional_marker: false,
-                                outer_sub_signature: None,
-                                code_signature: None,
-                                is_invocant: false,
-                                shape_constraints: None,
-                            })
-                            .collect()
-                    } else {
-                        let (use_positional, use_named) = Self::auto_signature_uses(&data.body);
-                        let mut defs = Vec::new();
-                        // A *bare* block `{ ... }` has an implicit `$_` parameter
-                        // (default from the outer topic). A pointy block `-> { ... }`
-                        // is also a `Block` (`is_bare_block` is set for the `.WHAT`),
-                        // but its signature is explicit — an empty `-> {}` takes no
-                        // arguments, so it must NOT gain the implicit `$_` (which
-                        // would render its signature as `($$_?)` and break
-                        // `.raku`/`.gist` round-tripping).
-                        let is_pointy = data
-                            .compiled_code
-                            .as_ref()
-                            .is_some_and(|cc| cc.is_pointy_block);
-                        if data.is_bare_block && !is_pointy && !use_positional {
-                            defs.push(ParamDef {
-                                name: "$_".to_string(),
-                                default: Some(Expr::Var("$_".to_string())),
-                                multi_invocant: true,
-                                required: false,
-                                named: false,
-                                slurpy: false,
-                                double_slurpy: false,
-                                onearg: false,
-                                sigilless: false,
-                                type_constraint: None,
-                                literal_value: None,
-                                sub_signature: None,
-                                where_constraint: None,
-                                traits: Vec::new(),
-                                optional_marker: true,
-                                outer_sub_signature: None,
-                                code_signature: None,
-                                is_invocant: false,
-                                shape_constraints: None,
-                            });
-                        }
-                        if use_positional {
-                            defs.push(ParamDef {
-                                name: "@_".to_string(),
-                                default: None,
-                                multi_invocant: true,
-                                required: false,
-                                named: false,
-                                slurpy: true,
-                                double_slurpy: false,
-                                onearg: false,
-                                sigilless: false,
-                                type_constraint: None,
-                                literal_value: None,
-                                sub_signature: None,
-                                where_constraint: None,
-                                traits: Vec::new(),
-                                optional_marker: false,
-                                outer_sub_signature: None,
-                                code_signature: None,
-                                is_invocant: false,
-                                shape_constraints: None,
-                            });
-                        }
-                        if use_named {
-                            defs.push(ParamDef {
-                                name: "%_".to_string(),
-                                default: None,
-                                multi_invocant: true,
-                                required: false,
-                                named: false,
-                                slurpy: true,
-                                double_slurpy: false,
-                                onearg: false,
-                                sigilless: false,
-                                type_constraint: None,
-                                literal_value: None,
-                                sub_signature: None,
-                                where_constraint: None,
-                                traits: Vec::new(),
-                                optional_marker: false,
-                                outer_sub_signature: None,
-                                code_signature: None,
-                                is_invocant: false,
-                                shape_constraints: None,
-                            });
-                        }
-                        defs
-                    }
-                });
-        let return_type = data.env.get("__mutsu_return_type").and_then(|v| match v {
-            Value::Str(s) => Some(s.to_string()),
-            _ => None,
-        });
-        let info = param_defs_to_sig_info(&param_defs, return_type);
-        // Build the owner sub key for parameter doc comment lookup.
-        // Must match the key format used by collect_doc_comments:
-        // - Subs use "&name" prefix
-        // - Methods use "ClassName::name" format
-        let sub_key = if !data.name.is_empty() {
-            let name = data.name.resolve();
-            // Check if the sub is a method (has a non-GLOBAL package context
-            // and uses Class::method format in doc comments)
-            let pkg = data.package.resolve();
-            if !pkg.is_empty() && pkg != "GLOBAL" {
-                Some(format!("{}::{}", pkg, name))
-            } else {
-                Some(format!("&{}", name))
-            }
-        } else {
-            None
-        };
-        make_signature_value_with_owner(info, sub_key)
-    }
-
-    pub(super) fn signature_required_positional_count(
-        info: &crate::value::signature::SigInfo,
-    ) -> i64 {
-        info.params
-            .iter()
-            .filter(|p| !p.named && !p.slurpy && !p.has_default && !p.optional_marker)
-            .count() as i64
-    }
-
-    fn signature_positional_count(info: &crate::value::signature::SigInfo) -> Option<i64> {
-        let mut count = 0i64;
-        for p in &info.params {
-            if p.named || (p.slurpy && p.sigil == '%') {
-                continue;
-            }
-            if p.slurpy {
-                return None;
-            }
-            count += 1;
-        }
-        Some(count)
-    }
-
-    pub(super) fn signature_count_value(info: &crate::value::signature::SigInfo) -> Value {
-        match Self::signature_positional_count(info) {
-            Some(count) => Value::Int(count),
-            None => Value::Num(f64::INFINITY),
-        }
-    }
-
-    pub(super) fn candidate_arity_value(infos: &[crate::value::signature::SigInfo]) -> Value {
-        let arity = infos
-            .iter()
-            .map(Self::signature_required_positional_count)
-            .min()
-            .unwrap_or(0);
-        Value::Int(arity)
-    }
-
-    pub(super) fn candidate_count_value(infos: &[crate::value::signature::SigInfo]) -> Value {
-        let mut max_count = 0i64;
-        for info in infos {
-            match Self::signature_positional_count(info) {
-                Some(count) => {
-                    if count > max_count {
-                        max_count = count;
-                    }
-                }
-                None => return Value::Num(f64::INFINITY),
-            }
-        }
-        Value::Int(max_count)
-    }
-
-    pub(super) fn shaped_dims_from_new_args(
-        &self,
-        args: &[Value],
-    ) -> Result<Option<Vec<usize>>, RuntimeError> {
-        let shape_val = match args.iter().find_map(|arg| match arg {
-            Value::Pair(name, value) if name == "shape" => Some(value.as_ref().clone()),
-            _ => None,
-        }) {
-            Some(v) => v,
-            None => return Ok(None),
-        };
-        let dims_vals = if let Some(items) = shape_val.as_list_items() {
-            items.to_vec()
-        } else {
-            match shape_val {
-                Value::Int(i) => vec![Value::Int(i)],
-                Value::Package(ref name) => {
-                    // Enum type as shape: use the number of enum variants
-                    if let Some(variants) = self.registry().enum_types.get(&name.resolve()) {
-                        vec![Value::Int(variants.len() as i64)]
-                    } else if name == "Bool" {
-                        // Bool is a builtin enum with 2 values (False, True)
-                        vec![Value::Int(2)]
-                    } else {
-                        return Ok(None);
-                    }
-                }
-                _ => return Ok(None),
-            }
-        };
-        let mut dims = Vec::with_capacity(dims_vals.len());
-        for dim in &dims_vals {
-            let n = match dim {
-                Value::Int(i) if *i > 0 => *i as usize,
-                Value::Int(i) => {
-                    return Err(RuntimeError::illegal_dimension_in_shape(*i));
-                }
-                Value::Num(f) if *f > 0.0 => *f as usize,
-                Value::Num(f) => {
-                    return Err(RuntimeError::illegal_dimension_in_shape(*f as i64));
-                }
-                Value::Package(name) => {
-                    if let Some(variants) = self.registry().enum_types.get(&name.resolve()) {
-                        variants.len()
-                    } else if name == "Bool" {
-                        2
-                    } else {
-                        return Ok(None);
-                    }
-                }
-                _ => return Ok(None),
-            };
-            dims.push(n);
-        }
-        if dims.is_empty() {
-            Ok(None)
-        } else {
-            Ok(Some(dims))
-        }
-    }
-
-    /// Build a (possibly multidimensional) shaped array of the given dims.
-    /// Each dimension allocates `dims[i]` slots via a fallible reservation, so
-    /// an absurd declared shape (`my @a[1e15]` = `Array.new(:shape(1e15))`)
-    /// yields a catchable `X::` instead of an uncatchable `handle_alloc_error`
-    /// abort. (raku aborts with a MoarVM panic on the same input.)
-    pub(super) fn make_shaped_array(dims: &[usize]) -> Result<Value, RuntimeError> {
-        if dims.is_empty() {
-            return Ok(Value::Nil);
-        }
-        let len = dims[0];
-        if dims.len() == 1 {
-            let mut items = Vec::new();
-            Self::autoviv_resize(&mut items, len, Value::Nil)?;
-            let value = Value::shaped_array(items);
-            crate::runtime::utils::mark_shaped_array(&value, Some(dims));
-            return Ok(value);
-        }
-        let child = Self::make_shaped_array(&dims[1..])?;
-        crate::runtime::utils::mark_shaped_array(&child, Some(&dims[1..]));
-        let mut items = Vec::new();
-        items.try_reserve(len).map_err(|_| {
-            RuntimeError::new(format!(
-                "Cannot allocate shaped array of {len} elements: memory allocation failed"
-            ))
-        })?;
-        items.extend((0..len).map(|_| child.clone()));
-        let value = Value::shaped_array(items);
-        crate::runtime::utils::mark_shaped_array(&value, Some(dims));
-        Ok(value)
-    }
-
-    pub(super) fn infer_array_shape(value: &Value) -> Option<Vec<usize>> {
-        if let Some(shape) = crate::runtime::utils::shaped_array_shape(value) {
-            return Some(shape);
-        }
-        let Value::Array(items, ..) = value else {
-            return None;
-        };
-        let mut shape = vec![items.len()];
-        let mut current = items.first().cloned();
-        while let Some(Value::Array(inner, ..)) = current {
-            shape.push(inner.len());
-            current = inner.first().cloned();
-        }
-        Some(shape)
-    }
-
-    pub(super) fn parse_parametric_type_name(name: &str) -> Option<(String, Vec<String>)> {
-        let base_end = name.find('[')?;
-        if !name.ends_with(']') {
-            return None;
-        }
-        let base = name[..base_end].trim().to_string();
-        let inner = &name[base_end + 1..name.len() - 1];
-        let args = inner
-            .split(',')
-            .map(|part| part.trim())
-            .filter(|part| !part.is_empty())
-            .map(ToOwned::to_owned)
-            .collect::<Vec<_>>();
-        Some((base, args))
-    }
-
-    pub(crate) fn call_method_all_with_values(
-        &mut self,
-        target: Value,
-        method: &str,
-        args: Vec<Value>,
-    ) -> Result<Vec<Value>, RuntimeError> {
-        if let Value::Instance {
-            class_name,
-            attributes,
-            id: target_id,
-        } = &target
-        {
-            if let Some(private_rest) = method.strip_prefix('!') {
-                // Resolve: owner-qualified (!Owner::method) or unqualified (!method)
-                let resolved = if let Some((owner_class, pm_name)) = private_rest.split_once("::") {
-                    self.resolve_private_method_with_owner(
-                        &class_name.resolve(),
-                        owner_class,
-                        pm_name,
-                        &args,
-                    )
-                    .map(|r| (r, pm_name))
-                } else {
-                    self.resolve_private_method_any_owner(
-                        &class_name.resolve(),
-                        private_rest,
-                        &args,
-                    )
-                    .map(|r| (r, private_rest))
-                };
-                if let Some(((resolved_owner, method_def), pm_name)) = resolved {
-                    let caller_class = self
-                        .method_class_stack
-                        .last()
-                        .cloned()
-                        .or_else(|| Some(self.current_package().to_string()));
-                    let caller_allowed = caller_class.as_deref() == Some(resolved_owner.as_str())
-                        || self
-                            .registry()
-                            .class_trusts
-                            .get(&resolved_owner)
-                            .is_some_and(|trusted| {
-                                caller_class
-                                    .as_ref()
-                                    .is_some_and(|caller| trusted.contains(caller))
-                            });
-                    if !caller_allowed {
-                        return Err(make_private_permission_error(
-                            pm_name,
-                            &class_name.resolve(),
-                            caller_class.as_deref().unwrap_or("GLOBAL"),
-                        ));
-                    }
-                    let (result, updated) = self.run_resolved_method_compiled_or_treewalk(
-                        &class_name.resolve(),
-                        &resolved_owner,
-                        pm_name,
-                        method_def,
-                        attributes.to_map(),
-                        args,
-                        Some(target.clone()),
-                    )?;
-                    attributes.commit_attrs(updated);
-                    return Ok(vec![result]);
-                } else {
-                    // Private method not found -- return structured X::Method::NotFound
-                    return Err(make_method_not_found_error(
-                        private_rest,
-                        &class_name.resolve(),
-                        true,
-                    ));
-                }
-            }
-
-            let candidates =
-                self.resolve_methods_per_mro_level(&class_name.resolve(), method, &args);
-            if !candidates.is_empty() {
-                let mut attrs = attributes.to_map();
-                let mut out = Vec::with_capacity(candidates.len());
-                for (resolved_owner, method_def) in candidates {
-                    // Build an invocant with the latest attributes so that
-                    // `$.attr` accessor reads inside the method body see
-                    // values updated by preceding calls in the MRO chain.
-                    let invocant = Value::write_back_sharing(
-                        attributes,
-                        *class_name,
-                        attrs.clone(),
-                        *target_id,
-                    );
-                    let (result, updated) = self.run_resolved_method_compiled_or_treewalk(
-                        &class_name.resolve(),
-                        &resolved_owner,
-                        method,
-                        method_def,
-                        attrs,
-                        args.clone(),
-                        Some(invocant),
-                    )?;
-                    attrs = updated;
-                    out.push(result);
-                }
-                attributes.commit_attrs(attrs);
-                return Ok(out);
-            }
-            // If the method is defined but no multi candidate matched,
-            // produce a dispatch error.
-            let own_class = class_name.resolve();
-            let method_exists = self.class_mro(&own_class).iter().any(|cn| {
-                self.registry()
-                    .classes
-                    .get(cn.as_str())
-                    .and_then(|c| c.methods.get(method))
-                    .is_some_and(|ovs| {
-                        // A submethod on an ancestor class is not visible to
-                        // the receiver, so don't count it as "method exists".
-                        let is_ancestor = cn.as_str() != own_class.as_str();
-                        ovs.iter()
-                            .any(|d| !d.is_private && (!d.is_my || !is_ancestor))
-                    })
-            });
-            if method_exists {
-                return Err(make_multi_no_match_error(method));
-            }
-        }
-
-        Ok(vec![self.call_method_with_values(target, method, args)?])
-    }
-
-    pub(crate) fn overwrite_array_items_by_identity_for_vm(
-        &mut self,
-        needle: &std::sync::Arc<crate::value::ArrayData>,
-        updated_items: Vec<Value>,
-        kind: ArrayKind,
-    ) {
-        self.overwrite_array_bindings_by_identity(
-            needle,
-            Value::Array(
-                std::sync::Arc::new(crate::value::ArrayData::new(updated_items)),
-                kind,
-            ),
-        );
     }
 }
