+++ src/runtime/mod.rs
@@ -195,6 +195,9 @@ mod methods_qualified;
 mod methods_quanthash_ctor;
 mod methods_seq_dispatch;
 pub(crate) mod methods_signature;
+mod methods_signature_candidates;
+pub(crate) mod methods_signature_errors;
+mod methods_signature_shaped;
 mod methods_string;
 mod methods_sub;
 mod methods_supply_dispatch;
